strtotime: improve
use long long: atleast 32-bit, but now time_t (real) to 32-bit or
64-bit is supported. Long long is C99 though, but that is fine.

check errno, it can have ERANGE.

+++ sfeed.c
@@ -529,7 +529,7 @@ printfields(void)
 	if (ctx.fields[FeedFieldTime].str.data)
 		r = parsetime(ctx.fields[FeedFieldTime].str.data, &t);
 	if (r != -1)
-		printf("%ld", (long)t);
+		printf("%lld", (long long)t);
 	putchar(FieldSeparator);
 	string_print_trimmed(&ctx.fields[FeedFieldTitle].str);
 	putchar(FieldSeparator);
